Cutting current event short
When drawing the currently active event, only the remaining duration of
the event will be drawn so nothing of this event will be shown before
the current time indicator.
Text might still be an issue. I added a minimum of degrees for a slice
to need to be able to get a title, but then, depending on the angle,
some text may still be shown across the current time indicator.
The text fadeout will probably help with this.